# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GenerateGeminiImageBody {
  prompt: string;
}

export interface GenerateGeminiImageResponse {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export type BakchoduMessageRole =
  (typeof BakchoduMessageRole)[keyof typeof BakchoduMessageRole];

export const BakchoduMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface BakchoduImageInput {
  /** Base64 encoded image data */
  data: string;
  /** MIME type of the image (e.g. image/jpeg, image/png, image/webp) */
  mimeType?: string;
}

export interface BakchoduMessage {
  role: BakchoduMessageRole;
  content: string;
  /** Images attached to this message */
  images?: BakchoduImageInput[];
}

export interface BakchoduChatRequest {
  message: string;
  history?: BakchoduMessage[];
  /** Optional images to analyze (base64 encoded) */
  images?: BakchoduImageInput[];
  /** GitHub or Railway token for authentication */
  token: string;
}

export interface BakchoduChatResponse {
  reply: string;
  model: string;
  tokensUsed?: number;
}

export interface BakchoduInfoResponse {
  name: string;
  creator: string;
  model: string;
  version: string;
  description: string;
  capabilities?: string[];
}