# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bakchodu/chat:
    post:
      operationId: bakchoduChat
      tags: [bakchodu]
      summary: Chat with BAKCHODU AI (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BakchoduChatRequest"
      responses:
        "200":
          description: BAKCHODU response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BakchoduChatResponse"
        "401":
          description: Unauthorized - invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /bakchodu/chat/stream:
    post:
      operationId: bakchoduChatStream
      tags: [bakchodu]
      summary: Stream chat with BAKCHODU AI (SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BakchoduChatRequest"
      responses:
        "200":
          description: SSE stream of BAKCHODU response
          content:
            text/event-stream: {}
        "401":
          description: Unauthorized - invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
...
components:
  schemas:
...
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BakchoduImageInput:
      type: object
      properties:
        data:
          type: string
          description: Base64 encoded image data
        mimeType:
          type: string
          description: MIME type of the image (e.g. image/jpeg, image/png, image/webp)
          default: image/jpeg
      required:
        - data
    BakchoduChatRequest:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/BakchoduMessage"
        images:
          type: array
          description: Optional images to analyze (base64 encoded)
          items:
            $ref: "#/components/schemas/BakchoduImageInput"
        token:
          type: string
          description: GitHub or Railway token for authentication
      required:
        - message
        - token
    BakchoduMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        images:
          type: array
          description: Images attached to this message
          items:
            $ref: "#/components/schemas/BakchoduImageInput"
      required:
        - role
        - content
    BakchoduChatResponse:
      type: object
      properties:
        reply:
          type: string
        model:
          type: string
        tokensUsed:
          type: integer
      required:
        - reply
        - model
...