# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate an image from a text prompt
 */
export const GenerateGeminiImageBody = zod.object({
  prompt: zod.string(),
});

export const GenerateGeminiImageResponse = zod.object({
  b64_json: zod.string(),
  mimeType: zod.string(),
});

const BakchoduImageInputSchema = zod.object({
  data: zod.string().describe("Base64 encoded image data"),
  mimeType: zod.string().optional().default("image/jpeg"),
});

/**
 * @summary Chat with BAKCHODU AI (non-streaming)
 */
export const BakchoduChatBody = zod.object({
  message: zod.string(),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
        images: zod.array(BakchoduImageInputSchema).optional(),
      }),
    )
    .optional(),
  images: zod.array(BakchoduImageInputSchema).optional(),
  token: zod.string().describe("GitHub or Railway token for authentication"),
});

export const BakchoduChatResponse = zod.object({
  reply: zod.string(),
  model: zod.string(),
  tokensUsed: zod.number().optional(),
});

/**
 * @summary Stream chat with BAKCHODU AI (SSE)
 */
export const BakchoduChatStreamBody = zod.object({
  message: zod.string(),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
        images: zod.array(BakchoduImageInputSchema).optional(),
      }),
    )
    .optional(),
  images: zod.array(BakchoduImageInputSchema).optional(),
  token: zod.string().describe("GitHub or Railway token for authentication"),
});

/**
 * @summary Get BAKCHODU server info
 */
export const BakchoduInfoResponse = zod.object({
  name: zod.string(),
  creator: zod.string(),
  model: zod.string(),
  version: zod.string(),
  description: zod.string(),
});